tpm2_checkquote.c: Fix untrusted loop bound
An attacker could control the number of times the loop iterates. In
parse_selection_data_from_file: An unscrutinized value from an untrusted source
used as a loop bound (CWE-606)

+++ tools/misc/tpm2_checkquote.c
@@ -273,7 +273,7 @@ static bool parse_selection_data_from_file(FILE *pcr_input,
         return false;
     }
 
-    UINT32 j;
+    size_t j;
     for (j = 0; j < le64toh(pcrs->count); j++) {
         if (fread(&pcrs->pcr_values[j], sizeof(TPML_DIGEST), 1, pcr_input)
                 != 1) {
